refactor(db): abstract verification operations into own module
